refactor: update style
