+++ app/components/layout/header.tsx
@@ -5,7 +5,7 @@ import { SearchBox } from "~/routes/_app/search-box";
 export function Header() {
 
   return (
-    <header className="sticky top-0 z-50 bg-white/80 dark:bg-gray-900/80 backdrop-blur-md border-b border-gray-200/50 dark:border-gray-700/50">
+    <header className="sticky top-0 z-50 bg-white/80 dark:bg-gray-900/80 backdrop-blur-md border-b border-gray-200/50 dark:border-gray-700/50 h-18">
       <div className="container mx-auto px-2 sm:px-4 py-4">
         <div className="flex justify-between items-center gap-1 sm:gap-2">
           <div className="flex-shrink-0">
